+++ src/module/_types.d.ts
@@ -4,13 +4,13 @@ import "./data/_types";
 import "./documents/_types";
 import {
   DrawSteelActor,
-  DrawSteelChatMessage
+  DrawSteelChatMessage,
 } from "./documents/_module.mjs";
 
 import Advancement from "./documents/advancement/advancement.mjs";
 import {
   PowerRoll,
-  ProjectRoll
+  ProjectRoll,
 } from "./rolls/_module.mjs";
 
 export interface AdvancementTypeConfiguration {
@@ -22,12 +22,12 @@ export interface AdvancementTypeConfiguration {
   /**
    * What item types this advancement can be used with.
    */
-  validItemTypes: Set < string > ;
+  validItemTypes: Set <string>;
 
   /**
    * Should this advancement type be hidden in the selection dialog?
    */
-  hidden ? : boolean;
+  hidden ?: boolean;
 }
 
 export interface PowerRollModifiers {
@@ -46,19 +46,19 @@ export interface RollPromptOptions {
   modifiers: PowerRollModifiers;
   formula: string;
   actor: DrawSteelActor;
-  data: Record < string, unknown > ;
-  skills: Set < string > ;
+  data: Record <string, unknown>;
+  skills: Set <string>;
 }
 
 export interface PowerRollPromptOptions extends RollPromptOptions {
   type: "ability" | "test";
   targets: PowerRollTargets[],
-  ability ? : string
+  ability ?: string
 }
 
 export interface PowerRollPrompt {
   rollMode: keyof typeof CONFIG["Dice"]["rollModes"];
-  powerRolls: Array < PowerRoll | DrawSteelChatMessage | object > ;
+  powerRolls: Array <PowerRoll | DrawSteelChatMessage | object>;
 }
 
 export interface ProjectRollPrompt {