+++ src/module/applications/api/_module.mjs
@@ -2,3 +2,6 @@ export { default as DSDocumentSheetMixin } from "./document-sheet-mixin.mjs";
 export { default as DSApplication } from "./application.mjs";
 export { default as DSDialog } from "./dialog.mjs";
 export { default as PseudoDocumentSheet } from "./pseudo-document-sheet.mjs";
+export { default as QueryManager } from "./query-manager.mjs";
+export { default as RollDialog } from "./roll-dialog.mjs";
+export { default as RollManager } from "./roll-manager.mjs";