+++ src/module/data/actor/base.mjs
@@ -1,5 +1,6 @@
 import DrawSteelChatMessage from "../../documents/chat-message.mjs";
 import { PowerRoll } from "../../rolls/power.mjs";
+import FormulaField from "../fields/formula-field.mjs";
 import { damageTypes, requiredInteger, setOptions } from "../helpers.mjs";
 import SizeModel from "../models/size.mjs";
 import SubtypeModelMixin from "../subtype-model-mixin.mjs";
@@ -16,7 +17,7 @@ const fields = foundry.data.fields;
 export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.TypeDataModel) {
   /** @inheritdoc */
   static defineSchema() {
-    const characteristic = { min: -5, max: 5, initial: 0, integer: true };
+    const characteristic = { min: -5, max: 5, initial: 0, integer: true, nullable: false };
     const schema = {};
 
     schema.stamina = new fields.SchemaField({
@@ -35,6 +36,10 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
     );
 
     schema.combat = new fields.SchemaField({
+      save: new fields.SchemaField({
+        threshold: new fields.NumberField({ required: true, nullable: false, integer: true, min: 1, max: 10, initial: 6 }),
+        bonus: new FormulaField(),
+      }),
       size: new fields.EmbeddedDataField(SizeModel),
       stability: requiredInteger({ initial: 0 }),
       turns: requiredInteger({ initial: 1 }),
@@ -43,9 +48,10 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
     schema.biography = new fields.SchemaField(this.actorBiography());
 
     schema.movement = new fields.SchemaField({
-      value: new fields.NumberField({ integer: true, min: 0, initial: 5 }),
+      value: new fields.NumberField({ nullable: false, integer: true, min: 0, initial: 5 }),
       types: new fields.SetField(setOptions(), { initial: ["walk"] }),
       hover: new fields.BooleanField(),
+      disengage: new fields.NumberField({ nullable: false, integer: true, min: 0, initial: 1 }),
     });
 
     schema.damage = new fields.SchemaField({
@@ -102,6 +108,9 @@ export default class BaseActorModel extends SubtypeModelMixin(foundry.abstract.T
     super.prepareDerivedData();
 
     this.stamina.winded = Math.floor(this.stamina.max / 2);
+
+    // Presents better if there's a 0 instead of blank
+    this.combat.save.bonus ||= "0";
 
     const highestCharacteristic = Math.max(0, ...Object.values(this.characteristics).map(c => c.value));