+++ src/module/data/subtype-model-mixin.mjs
@@ -2,7 +2,8 @@
 
 /**
  * Mixin for common functions used across most or all document subtypes.
- * @mixin
+ * @template {import("@common/_types.mjs").Constructor<foundry.abstract.TypeDataModel>} ModelClass
+ * @param {ModelClass} base
  */
 export default base => {
   return class DrawSteelSystemModel extends base {